Feat check whether request client is writer that article or not
- 삭제 요청이 들어온 게시글의 작성자가 request한 요청자의 게시글인지 확인한 후 삭제하도록 변경.
- addPost의 status(200)에서 201로 변경(HTTP 표준 스펙)

+++ services/freeBoard.js
@@ -26,9 +26,15 @@ const setPost = async (req, res, next) => {
     // const { id, categoryId, title, content } = req.body;
     // const existingPost = await freeBoardRepos.checkDeletedPost(id);
     const existingPost = await boardRepo.checkPost(id, model);
+
     if (!existingPost) {
       throw new Error("이미 삭제된 공고입니다.");
     }
+
+    if (req.user.id !== existingPost.UserId) {
+      throw new Error("글 작성자가 아닙니다.");
+    }
+
     const updatedPost = await boardRepo.updatePost(id, title, content, model);
     // const updatedPost = await freeBoardRepos.updatePost(
     //   id,
@@ -51,10 +57,19 @@ const setPost = async (req, res, next) => {
 const deletePost = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const deletedPost = await boardRepo.destroyPost(id, model);
-    if (!deletedPost) {
+
+    const existingPost = await boardRepo.checkPost(id, model);
+
+    if (!existingPost) {
       throw new Error("이미 삭제된 공고입니다.");
     }
+
+    if (req.user.id !== existingPost.UserId) {
+      throw new Error("글 작성자가 아닙니다.");
+    }
+
+    await boardRepo.destroyPost(id, model);
+
     return res.status(200).json({ message: "Posting is deleted" });
   } catch (err) {
     next(err);
@@ -63,12 +78,13 @@ const deletePost = async (req, res, next) => {
 
 const addPost = async (req, res, next) => {
   try {
-    const userId = req.userId;
+    const userId = req.user.id;
     const { title, content } = req.body;
     // const { userId, categoryId, title, content } = req.body;
+    console.log(userId);
     await boardRepo.createPost(title, content, userId, model);
     // await freeBoardRepos.createPost(userId, categoryId, title, content);
-    return res.status(200).json({ message: "jobPosting is created" });
+    return res.status(201).json({ message: "jobPosting is created" });
   } catch (err) {
     next(err);
   }
